+++ app/Http/Controllers/ProfileController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
-use App\User;
+use App\Models\User;
 use Auth;
 use Image;
 use File;
