+++ app/Models/Article.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -41,7 +41,7 @@ class Article extends Model
 	{
 		return static::latest()
 			->paginate();
-	} 
+	}
 
 	public static function loadAllMine($user_id)
 	{
