+++ app/Models/VerifyUser.php
@@ -1,15 +1,16 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use App\Models\User;
 
 class VerifyUser extends Model
 {
     protected $guarded = [];
     protected $fillable= ['user_id','verify_token'];
-    
+
     public function user(){
-        return $this->belongsTo("App\User");
+        return $this->belongsTo(User::class);
     }
 }
